# storage: assert iterator invariants