Resolves #86651.

**storage: remove `Valid()` call in `pebbleIterator.UnsafeKey()`**

This patch removes an unnecessary `Valid()` call in
`pebbleIterator.UnsafeKey()`. The `SimpleMVCCIterator` API contract
already mandates that the caller must call `Valid()` before calling
`UnsafeKey()`.

This is done in preparation for iterator invariant checks. These will be
performed in `Valid()` and this call would cause infinite recursion.
This additionally gives a nice performance boost (using the
`SSTIterator` benchmark since `MVCCScan` does not call `UnsafeKey`):

```
name                                                   old time/op  new time/op  delta
SSTIterator/keys=10000/variant=pebble/verify=false-24  1.08ms ± 0%  1.05ms ± 0%  -2.51%  (p=0.000 n=10+8)
```

Release note: None
  
**storage: add `pebbleIterator` invariant assertions**

This patch adds invariant checks for `pebbleIterator`. General API
invariant checks for the `SimpleMVCCIterator` and `MVCCIterator`
interfaces are also added, and applied to `pebbleIterator`.

Release note: None
  
**storage: add `intentInterleavingIter` invariant assertions**

This patch adds invariant assertions for `intentInterleavingIter` using
a similar pattern as `pointSynthesizingIter`: by calling
`assertInvariants()` under race during `Valid()`. This allows specifying
all internal invariants in a simple-to-understand form and asserting
them all after every state transition.

Existing race assertions have been migrated.

Release note: None
  
**storage: fix `pointSynthesizingIter.Key()` and `UnsafeRawKey()`**

`pointSynthesizingIter.Key()` could return a different key than
`UnsafeKey()`. Additionally, `UnsafeRawKey()` would omit the key's
timestamp. Luckily, these methods have no current callers.

Release note: None

**storage: add `pointSynthesizingIter` invariant assertions**

Release note: None

**storage: add `MVCCIncrementalIterator` invariant assertions**

Release note: None
  
**storage: add `ReadAsOfIterator` invariant assertions**

Release note: None