@erikgrinaker 2022-09-13
This should be ready to review now. Found a couple of bugs, and a bonus performance optimization as well.